backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.30.6
pydantic>=2.6,<3
pydantic-settings>=2.2,<3
sqlalchemy==2.0.34
alembic==1.13.2
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.27.2

# RAG (optional but recommended)
sentence-transformers==3.0.1
faiss-cpu==1.13.2
numpy==1.26.4
passlib[argon2]